Add dismiss task endpoint
